refactor: address final async API PR feedback
- Moved JobState and JobStatus types from public crate::types to crate::api::types
- Removed public types/events.rs module to prevent Alef binding generation
- Restored OCTET_STREAM_MIME_TYPE to pub(crate) visibility
- Added concurrent uniqueness test for job ID generation
- Verified clean build and integration test coverage
